requirements.txt
alignn==2023.10.1
ase==3.22.1
brokenaxes==0.5.0
dgl==1.0.1+cu117
fasteners==0.19
jarvis_tools==2023.12.12
matplotlib==3.8.2
mp_api==0.39.4
numpy==1.26.3
pandas==2.1.4
pymatgen==2023.12.18
pytorch_ignite==0.5.0.dev20240103
pytorch_ignite==0.4.8
pytorch_lightning==2.1.3
pytorch_lightning==1.8.6
Requests==2.31.0
scikit_learn==1.3.2
setuptools==69.0.3
torch==2.0.0
torch==1.13.1.post200
torchmetrics==1.0.1
torchmetrics==0.11.0
tqdm==4.66.1
